Fix some compile issues.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5339 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/web/gl_vidweb.c
@@ -380,17 +380,17 @@ void INS_EnumerateDevices(void *ctx, void(*callback)(void *ctx, const char *type
 	char foobar[64];
 	for (i = 0; i < countof(gamepaddeviceids); i++)
 	{
-		Q_snprintfz(foobar, sizeof(foobar), "gp%i", i);
+		Q_snprintfz(foobar, sizeof(foobar), "gp%i", (int)i);
 		callback(ctx, "gamepad", foobar, &gamepaddeviceids[i]);
 	}
 	for (i = 0; i < countof(mouseid); i++)
 	{
-		Q_snprintfz(foobar, sizeof(foobar), "m%i", i);
+		Q_snprintfz(foobar, sizeof(foobar), "m%i", (int)i);
 		callback(ctx, "mouse", foobar, &mouseid[i]);
 	}
 	for (i = 0; i < countof(keyboardid); i++)
 	{
-		Q_snprintfz(foobar, sizeof(foobar), "kb%i", i);
+		Q_snprintfz(foobar, sizeof(foobar), "kb%i", (int)i);
 		callback(ctx, "keyboard", foobar, &keyboardid[i]);
 	}
 }
